requirements.txt
Flask==3.0.0
requests==2.31.0
mysql-connector-python==8.1.0
pandas==2.1.1
numpy==1.24.3
plotly==5.18.0